+++ packages/core/src/physics/Collision.ts
@@ -1,20 +1,6 @@
-import { Vector3 } from "@galacean/engine-math";
+import { ContactPoint } from "./ContactPoint";
 import { ColliderShape } from "./shape";
 import { ICollision } from "@galacean/engine-design";
-
-/**
- * Describes a contact point where the collision occurs.
- */
-export class ContactPoint {
-  /** The position of the contact point between the shapes, in world space. */
-  readonly position = new Vector3();
-  /** The normal of the contacting surfaces at the contact point. The normal direction points from the second shape to the first shape. */
-  readonly normal = new Vector3();
-  /** The impulse applied at the contact point, in world space. Divide by the simulation time step to get a force value. */
-  readonly impulse = new Vector3();
-  /** The separation of the shapes at the contact point.  A negative separation denotes a penetration. */
-  separation: number;
-}
 
 /**
  * Collision information between two shapes when they collide.
@@ -42,21 +28,18 @@ export class Collision {
    */
   getContacts(outContacts: ContactPoint[]): number {
     const { shape0Id, shape1Id } = this._nativeCollision;
+    const factor = shape0Id < shape1Id ? 1 : -1;
+
     const nativeContactPoints = this._nativeCollision.getContacts();
     const length = nativeContactPoints.size();
-    for (let i = 0, n = length; i < n; i++) {
+    for (let i = 0; i < length; i++) {
       const nativeContractPoint = nativeContactPoints.get(i);
-      const { position, normal, impulse, separation } = nativeContractPoint;
-      let factor = 1;
-      if (shape0Id > shape1Id) {
-        factor = -1;
-      }
 
       const contact = (outContacts[i] ||= new ContactPoint());
-      contact.position.set(position.x, position.y, position.z);
-      contact.normal.set(normal.x, normal.y, normal.z).scale(factor);
-      contact.impulse.set(impulse.x, impulse.y, impulse.z).scale(factor);
-      contact.separation = separation;
+      contact.position.copyFrom(nativeContractPoint.position);
+      contact.normal.copyFrom(nativeContractPoint.normal).scale(factor);
+      contact.impulse.copyFrom(nativeContractPoint.impulse).scale(factor);
+      contact.separation = nativeContractPoint.separation;
     }
     return length;
   }
